Extract rich text processing into a separate file + add image directory processing + complete test for document generation with embedded rich text
