+++ source/docx_generator/utils.py
@@ -1,6 +1,3 @@
-import uuid
-from typing import Callable
-
 from docx.enum.text import WD_PARAGRAPH_ALIGNMENT
 
 
@@ -13,7 +10,3 @@ def resize_image(image, new_width):
 
 def get_available_paragraph_alignments():
     return list(WD_PARAGRAPH_ALIGNMENT.__members__.keys())
-
-
-def generate_logger_identifier():
-    return uuid.uuid4().hex[:10]
