+++ test/component/test_docx_generator.py
@@ -20,12 +20,15 @@
 import json
 import os
 import re
+import shutil
 from unittest import TestCase
 
 from docx import Document
-from docx.enum.style import WD_STYLE_TYPE
+from docx.drawing import Drawing
+from docx.enum.text import WD_PARAGRAPH_ALIGNMENT, WD_COLOR_INDEX
 from docx.opc.constants import RELATIONSHIP_TYPE
-from docx.styles.styles import Styles
+from docx.text.paragraph import Paragraph
+from docx.text.run import Run
 
 from docx_generator.docx_generator import DocxGenerator
 from docx_generator.exceptions.rendering_error import RenderingError
@@ -57,6 +60,7 @@ class TestDocxGenerator(TestCase):
         self._subject = DocxGenerator(logger_mode='DEBUG')
 
     def tearDown(self) -> None:
+        shutil.rmtree(os.path.join(self._base_path, 'tmp', 'images'))
         return None
         for filename in self._output_filenames.values():
             try:
@@ -303,12 +307,16 @@ class TestDocxGenerator(TestCase):
 
     def test_should_generate_docx_with_richtext(self):
         richtext_content = [
+            # 0
+            # HEADING 1
             {
                 "type": "heading-1",
                 "children": [
                     {"text": "This is a test for Richtext"}
                 ]
             },
+            # 1
+            # PARAGRAPH 1
             {
                 "type": "paragraph",
                 "children": [
@@ -317,7 +325,11 @@ class TestDocxGenerator(TestCase):
                     {"text": "with italic and underline", "italic": True, "underline": True}
                 ]
             },
+            # 2
+            # PARAGRAPH 2
             {"type": "paragraph", "children": []},
+            # 3
+            # PARAGRAPH 3
             {
                 "type": "paragraph",
                 "align": "center",
@@ -325,21 +337,45 @@ class TestDocxGenerator(TestCase):
                     {"text": "code example Lli", "code": True}
                 ]
             },
-            {"type": "caption", "text": "This is the figure caption 1"},
+            # 4
+            # CAPTION 1
+            {
+                "type": "caption",
+                "align": "center",
+                "text": "This is the figure caption 1"
+            },
+            # 5
+            # PARAGRAPH 4
             {"type": "paragraph", "children": []},
+            # 6
+            # IMAGE 1
             {"type": "image-uuid", "image_uuid": "5bacc2bc-5b90-4c47-93d7-d9291911c4b3"},
+            # 7
+            # BLOCK QUOTE 1
             {"type": "block-quote", "children": [{"text": "This is a test for Block Quote"}]},
+            # 8
+            # BLOCK QUOTE 2
             {"type": "block-quote", "children": [{"text": "With multiple lines"}]},
+            # 9
+            # BULLETED LIST 1
             {"type": "bulleted-list", "children": [
                 {"type": "list-item", "children": [{"text": "Item1"}]},
-                {"type": "list-item", "children": [{"text": "Item1.2"}]}
+                {"type": "list-item", "children": [{"text": "Item2"}]}
             ]},
+            # 10
+            # PARAGRAPH 5
             {"type": "paragraph", "children": []},
+            # 11
+            # NUMBERED LIST 1
             {"type": "numbered-list", "children": [
                 {"type": "list-item", "children": [{"text": "AAA"}]},
                 {"type": "list-item", "children": [{"text": "BBB"}]}
             ]},
+            # 12
+            # PARAGRAPH 6
             {"type": "paragraph", "children": []},
+            # 13
+            # PARAGRAPH 7
             {
                 "type": "paragraph",
                 "align": "center",
@@ -347,12 +383,20 @@ class TestDocxGenerator(TestCase):
                     {"text": "Separation between two lists"}
                 ]
             },
+            # 14
+            # PARAGRAPH 8
             {"type": "paragraph", "children": []},
+            # 15
+            # NUMBERED LIST 2
             {"type": "numbered-list", "children": [
                 {"type": "list-item", "children": [{"text": "CCC"}]},
                 {"type": "list-item", "children": [{"text": "DDD"}]}
             ]},
+            # 16
+            # PARAGRAPH 9
             {"type": "paragraph", "children": []},
+            # 17
+            # TABLE
             {
                 "type": "table",
                 "children": [
@@ -364,6 +408,7 @@ class TestDocxGenerator(TestCase):
                                 "children": [
                                     {
                                         "type": "paragraph",
+                                        "align": "right",
                                         "children": [
                                             {
                                                 "text": "aa"
@@ -406,39 +451,33 @@ class TestDocxGenerator(TestCase):
                             {
                                 "type": "table-cell",
                                 "children": [
-                                    {
-                                        "type": "paragraph",
-                                        "children": [
-                                            {
-                                                "text": "dd"
-                                            }
-                                        ]
-                                    }
+                                    {"type": "bulleted-list", "children": [
+                                        {"type": "list-item", "children": [{"text": "Item in cell 1"}]},
+                                        {"type": "list-item", "children": [{"text": "Item in cell 2"}]}
+                                    ]}
                                 ]
                             }
                         ]
                     }
                 ]
             },
+            # 18
+            # CAPTION 2
             {"type": "caption", "text": "This is the figure caption 2"},
+            # 19
+            # PARAGRAPH 10
+            {"type": "paragraph", "children": []},
+            # 20
+            # IMAGE 2
+            {"type": "image", "image_path": "https://www.vivelapub.fr/wp-content/uploads/2012/01/mire-tdf.jpg"},
+            {"type": "image", "image_path": "/etc/passwd"}
         ]
 
         data = {
             'value': json.dumps(richtext_content)
         }
 
-        print(json.dumps(richtext_content))
-
-        styles = self._subject.get_available_styles(self._base_path, os.path.join(self._template_path, 'richtext_global_template.docx'))
-        # for style in styles:
-        #     print(style.get('name'), style.get('id'))
-
-        self._subject.generate_docx(
-            self._base_path,
-            os.path.join(self._template_path, 'richtext_global_template.docx'),
-            data,
-            os.path.join(self._results_path, self._output_filenames['richtext_result']),
-            {
+        style_mapping = {
                 "heading-1": "Heading1",
                 "heading-2": "Heading2",
                 "heading-3": "Heading3",
@@ -451,8 +490,158 @@ class TestDocxGenerator(TestCase):
                 "bulleted-list": "APBulleted",
                 "caption": "Caption"
             }
+
+        self._subject.generate_docx(
+            self._base_path,
+            os.path.join(self._template_path, 'richtext_global_template.docx'),
+            data,
+            os.path.join(self._results_path, self._output_filenames['richtext_result']),
+            style_mapping
         )
 
-        result_file = Document(os.path.join(self._results_path, self._output_filenames['richtext_result']))
+        result_file = Document(os.path.join(self._results_path, self._output_filenames['richtext_result']))  # type: Document
+
+        result_file_paragraphs = result_file.paragraphs
+
+        self.assertEqual(28, len(result_file_paragraphs))
+
+        # HEADING 1
+        h1_paragraph = result_file_paragraphs[0]  # type: Paragraph
+        self.assertEqual('Heading 1', h1_paragraph.style.name)
+        self.assertEqual(richtext_content[0].get('children', [])[0].get('text'), h1_paragraph.text)
+
+        # PARAGRAPH 1
+        paragraph1 = result_file_paragraphs[1]  # type: Paragraph
+        paragraph1_r1 = paragraph1.runs[0]
+        self.assertEqual(richtext_content[1].get('children', [])[0].get('text'), paragraph1_r1.text)
+        paragraph1_r2 = paragraph1.runs[1]
+        self.assertEqual(richtext_content[1].get('children', [])[1].get('text'), paragraph1_r2.text)
+        self.assertEqual(True, paragraph1_r2.bold)
+        paragraph1_r3 = paragraph1.runs[2]
+        self.assertEqual(richtext_content[1].get('children', [])[2].get('text'), paragraph1_r3.text)
+        self.assertEqual(True, paragraph1_r3.italic)
+        self.assertEqual(True, paragraph1_r3.underline)
+
+        # PARAGRAPH 2
+        paragraph2 = result_file_paragraphs[2]  # type: Paragraph
+        self.assertEqual('', paragraph2.text)
+
+        # PARAGRAPH 3
+        paragraph3 = result_file_paragraphs[3]  # type: Paragraph
+        self.assertEqual(WD_PARAGRAPH_ALIGNMENT.CENTER, paragraph3.alignment)
+        paragraph3_r1 = paragraph3.runs[0]
+        self.assertEqual(richtext_content[3].get('children', [])[0].get('text'), paragraph3_r1.text)
+        self.assertEqual('Courier New', paragraph3_r1.font.name)
+        self.assertEqual(WD_COLOR_INDEX.GRAY_25, paragraph3_r1.font.highlight_color)
+
+        # CAPTION 1
+        caption1 = result_file_paragraphs[4]  # type: Paragraph
+        self.assertEqual(WD_PARAGRAPH_ALIGNMENT.CENTER, caption1.alignment)
+        self.assertEqual(4, len(caption1.runs))
+        caption1_r1 = caption1.runs[0]
+        self.assertEqual('Figure ', caption1_r1.text)
+        caption1_r2 = caption1.runs[1]  # type: Run
+        # TODO: Test if caption numbering exist
+        caption1_r3 = caption1.runs[2]  # type: Run
+        self.assertEqual(' ', caption1_r3.text)
+        caption1_r4 = caption1.runs[3]  # type: Run
+        self.assertEqual(richtext_content[4].get('text'), caption1_r4.text)
+
+        # PARAGRAPH 4
+        paragraph2 = result_file_paragraphs[5]  # type: Paragraph
+        self.assertEqual('', paragraph2.text)
+
+        # IMAGE 1
+        image1 = result_file_paragraphs[6]  # type: Paragraph
+        image1_r1 = image1.runs[0]  # type: Run
+        self.assertEqual(1, len(list(image1_r1.iter_inner_content())))
+        self.assertEqual(Drawing, type(list(image1_r1.iter_inner_content())[0]))
+
+        # BLOCK QUOTE 1
+        block_quote1 = result_file_paragraphs[7]  # type: Paragraph
+        self.assertEqual(style_mapping['block-quote'], block_quote1.style.style_id)
+        self.assertEqual(richtext_content[7].get('children', [])[0].get('text'), block_quote1.text)
+
+        # BLOCK QUOTE 2
+        block_quote1 = result_file_paragraphs[8]  # type: Paragraph
+        self.assertEqual(style_mapping['block-quote'], block_quote1.style.style_id)
+        self.assertEqual(richtext_content[8].get('children', [])[0].get('text'), block_quote1.text)
+
+        # BULLETED LIST 1
+        bulleted_list1_item1 = result_file_paragraphs[9]  # type: Paragraph
+        self.assertEqual(style_mapping['bulleted-list'], bulleted_list1_item1.style.style_id)
+        self.assertEqual(richtext_content[9].get('children', [])[0].get('children', [])[0].get('text'), bulleted_list1_item1.text)
+
+        bulleted_list1_item2 = result_file_paragraphs[10]  # type: Paragraph
+        self.assertEqual(style_mapping['bulleted-list'], bulleted_list1_item2.style.style_id)
+        self.assertEqual(richtext_content[9].get('children', [])[1].get('children', [])[0].get('text'), bulleted_list1_item2.text)
+
+        # PARAGRAPH 5
+        paragraph2 = result_file_paragraphs[11]  # type: Paragraph
+        self.assertEqual('', paragraph2.text)
+
+        # NUMBERED LIST 1
+        numbered_list1_item1 = result_file_paragraphs[12]  # type: Paragraph
+        self.assertEqual(style_mapping['numbered-list'], numbered_list1_item1.style.style_id)
+        self.assertEqual(richtext_content[11].get('children', [])[0].get('children', [])[0].get('text'), numbered_list1_item1.text)
+
+        numbered_list1_item2 = result_file_paragraphs[13]  # type: Paragraph
+        self.assertEqual(style_mapping['numbered-list'], numbered_list1_item2.style.style_id)
+        self.assertEqual(richtext_content[11].get('children', [])[1].get('children', [])[0].get('text'), numbered_list1_item2.text)
+
+        # PARAGRAPH 6
+        paragraph6 = result_file_paragraphs[14]  # type: Paragraph
+        self.assertEqual('', paragraph6.text)
+
+        # PARAGRAPH 7
+        paragraph7 = result_file_paragraphs[15]  # type: Paragraph
+        self.assertEqual(WD_PARAGRAPH_ALIGNMENT.CENTER, paragraph7.alignment)
+        self.assertEqual('Separation between two lists', paragraph7.text)
+
+        # PARAGRAPH 8
+        paragraph8 = result_file_paragraphs[16]  # type: Paragraph
+        self.assertEqual('', paragraph8.text)
+
+        # NUMBERED LIST 2
+        numbered_list2_item1 = result_file_paragraphs[17]  # type: Paragraph
+        self.assertEqual(style_mapping['numbered-list'], numbered_list2_item1.style.style_id)
+        self.assertEqual(richtext_content[15].get('children', [])[0].get('children', [])[0].get('text'), numbered_list2_item1.text)
+
+        numbered_list2_item2 = result_file_paragraphs[18]  # type: Paragraph
+        self.assertEqual(style_mapping['numbered-list'], numbered_list2_item2.style.style_id)
+        self.assertEqual(richtext_content[15].get('children', [])[1].get('children', [])[0].get('text'), numbered_list2_item2.text)
+
+        # PARAGRAPH 9
+        paragraph9 = result_file_paragraphs[19]  # type: Paragraph
+        self.assertEqual('', paragraph9.text)
+
+        # TODO: test table
+        # TABLE
+        table = result_file.tables
+
+        # CAPTION 2
+        caption2 = result_file_paragraphs[20]  # type: Paragraph
+        self.assertEqual(WD_PARAGRAPH_ALIGNMENT.CENTER, caption1.alignment)
+        self.assertEqual(4, len(caption1.runs))
+        caption2_r1 = caption2.runs[0]
+        self.assertEqual('Figure ', caption2_r1.text)
+        caption2_r2 = caption2.runs[1]  # type: Run
+        # TODO: Test if caption numbering exist
+        caption2_r3 = caption2.runs[2]  # type: Run
+        self.assertEqual(' ', caption2_r3.text)
+        caption3_r4 = caption2.runs[3]  # type: Run
+        self.assertEqual(richtext_content[18].get('text'), caption3_r4.text)
+
+        # PARAGRAPH 10
+        paragraph10 = result_file_paragraphs[21]  # type: Paragraph
+        self.assertEqual('', paragraph10.text)
+
+        # IMAGE 2
+        image2 = result_file_paragraphs[21]  # type: Paragraph
+        image2_r1 = image2.runs[0]  # type: Run
+        self.assertEqual(1, len(list(image2_r1.iter_inner_content())))
+        self.assertEqual(Drawing, type(list(image2_r1.iter_inner_content())[0]))
+
+        # IMAGE 3
 
         self.assertTrue(False)
